Retrieve Azure user-id by any mail address

+++ contrib/set_teams_presence.sh
@@ -10,20 +10,19 @@ client_secret='<M365 client ID>'
 scope='https://graph.microsoft.com/.default'
 grant_type='client_credentials'
 
-#status='{"sessionId":"'${client_id}'","availability":"Available","activity":"Available","expirationDuration":"PT1H"}'
-#status='{"sessionId":"'${client_id}'","availability":"Busy","activity":"InACall","expirationDuration":"PT2H"}'
 status='{"sessionId":"'${client_id}'"}'
 
 token_file="./tmp/azure.client_token"
 response="./tmp/azure.response"
+user_file="./tmp/azure.users"
 
 if [ "$#" -lt 2 ] ; then
-	echo "Usage: $0 <user> <available|busy"
+	echo "Usage: $0 <user> <Available|Busy"
 	exit 1
 elif [ -n "${1//[-A-Za-z0-9@._]}" ] ; then
 	echo "Error: <user> can only contain alpha-numeric characters and any of '-_.@'"
 	exit 1
-elif [ "$2" = "busy" ] ; then
+elif [ "$2" = "Busy" ] ; then
 	status='{"sessionId":"'${client_id}'","availability":"Busy","activity":"InACall","expirationDuration":"PT2H"}'
 fi
 
@@ -35,9 +34,13 @@ update_token()
 	 --data-urlencode "client_id=${client_id}" \
 	 --data-urlencode "client_secret=${client_secret}" \
 	 --data-urlencode "scope=${scope}" \
-	 --data-urlencode "grant_type=${grant_type}" | jq -r '.access_token')
+	 --data-urlencode "grant_type=${grant_type}" | jq -r '.access_token|values')
 
-	echo $token > "${token_file}"
+	if [ -z "${token}" ] ; then
+		echo "Error: something went wrong"
+		exit 1
+	fi
+	echo "${token}" > "${token_file}"
 	chmod 600 "${token_file}"
 }
 
@@ -52,27 +55,37 @@ get_token()
 
 get_user()
 {
-	error=$(curl -s -X GET -H "Authorization: Bearer ${token}" -w "%{http_code}" -o "${response}" "https://graph.microsoft.com/v1.0/users/${user}")
+	# check cached id first
+	id=$(cat "${user_file}" | grep -m 1 "^${user}" | cut -d" " -f2)
+	if [ -z "$id" ] ; then
+		error=$(curl -s -X GET -H "Authorization: Bearer ${token}" -w "%{http_code}" -o "${response}" "https://graph.microsoft.com/v1.0/users?\$filter=proxyAddresses/any(c:c+eq+'smtp:${user}')")
+
+		if [ "$error" = "401" ] ; then
+			update_token
+			get_user
+		fi
+
+		if [ "$error" = "404" ] ; then
+			echo "Error: <user> (${user}) not found!"
+			exit 1
+		elif [ "$error" != "200" ] ; then
+			echo "Error: unhandled HTTP ${error}"
+			exit 1
+		fi
+
+		id=$(jq -r '.value[0].id|values' "${response}")
+		rm -f "${response}"
+		if [ -n "${id}" ] ; then
+			echo "${user} ${id}" >> "${user_file}"
+		else
+			echo "Error: no such user ${user}"
+			exit 1
+		fi
+	fi
 }
 
 get_token
 get_user
-
-if [ "$error" = "401" ] ; then
-	update_token
-	get_user
-fi
-
-if [ "$error" = "404" ] ; then
-	echo "Error: <user> (${user}) not found!"
-	exit 1
-elif [ "$error" != "200" ] ; then
-	echo "Error: unhandled HTTP ${error}"
-	exit 1
-fi
-
-id=$(jq -r '.id' "${response}")
-rm -f "${response}"
 
 if [ "${status//availability}" != "${status}" ] ; then
 	error=$(curl -s -X POST -w "%{http_code}" -o "${response}" -H "Authorization: Bearer ${token}" -H "Content-Type: application/json" "https://graph.microsoft.com/v1.0/users/${id}/presence/setPresence" -d "${status}")
@@ -82,6 +95,6 @@ fi
 
 if [ "$error" != "200" ] ; then
 	echo "Error: unhandled HTTP ${error}"
-	cat ${response}
+	cat "${response}"
 fi
 rm -f "${response}"
